+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0-test5] - 2026-02-23
+
+### Changed
+- **Inclusions / Exclusions:**
+  - Changed workflow for adding inclusions / exclusions to have both buttons instead of first clicking "Add".
+  - Made hidden settings show hidden icon in orange. 
+  - Added note that this tool is an advanced feature with clickable link to journal. 
+  - Added "All" filter for namespace list to show all settigs (this may load slowly).
+
 ## [0.7.0-test4] - 2026-02-22
 
 ### Fixed